% Generated by roxygen2 (4.0.2): do not edit by hand
\name{getCountsFc}
\alias{getCountsFc}
\title{Get counts using featureCounts}
\usage{
getCountsFc(samp, bsgenome = NULL, ranges = NULL, chrs = NULL, fragsize,
  winsize = 50, ncore = 1)
}
\arguments{
\item{samp}{Sample data.frame from readSampleInfo()}

\item{bsgenome}{}

\item{chrs}{}

\item{fragsize}{Average fragment length from the sequencing experiment. Reads will be extended up to this size when computing coverage.}
}
\value{
desc
}
\description{
desc
}

