% Generated by roxygen2 (4.0.2): do not edit by hand
\name{getReads}
\alias{getReads}
\title{Store GenomicRanges of BAM reads in a list}
\usage{
getReads(samp, bsgenome, chrs, fragsize, ncore)
}
\arguments{
\item{samp}{Sample data.frame from readSampleInfo()}

\item{bsgenome}{}

\item{chrs}{}

\item{fragsize}{Average fragment length from the sequencing experiment. Reads will be extended up to this size when computing coverage.}
}
\value{
A list of GenomicRanges objects.
}
\description{
Will return a list of GenomicRanges objects
}

