% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/reporting.r
\name{maBed}
\alias{maBed}
\title{Write BED file of DMR regions}
\usage{
maBed(ma, file)
}
\arguments{
\item{ma}{Output list from a run of methylaction()}

\item{file}{Name of BED file to create}
}
\value{
Writes BED file to disk.
}
\description{
Creates a BED file suitable for uploading as a custom track to the UCSC genome browser.
}

