% Generated by roxygen2 (4.0.1): do not edit by hand
\name{maClustering}
\alias{maClustering}
\title{Clustering of samples based on read counts}
\usage{
maClustering(samp, ma, mincv, type, pdf)
}
\arguments{
\item{samp}{Description of samples from readSampleInfo()}

\item{ma}{Output object from methylaction()}

\item{mincv}{Minimum cv (cv=mean/sd) for a window to be included in the clustering.}

\item{type}{One of "mds" or "hca"}

\item{pdf}{PDF file to output.}
}
\value{
Saves plot to disk.
}
\description{
Will plot clustering based on some subset of read counts.
}

