% Generated by roxygen2 (4.0.2): do not edit by hand
\name{maCompare}
\alias{maCompare}
\title{Compare between various runs of methylaction()}
\usage{
maCompare(malist)
}
\arguments{
\item{malist}{List of output objects from methylaction(). Set the names() attribute of this list to unique descriptive names for each run.}
}
\value{
A list containing both site by site comparisions and a data.frame summary table of total and shared sites for each pairwise comparision between runs.
}
\description{
Reports shared and non-shared regions detected by different runs of methylaction(). Useful for comparing between studies/cohorts and paramater tuning.
}

