% Generated by roxygen2 (4.0.1): do not edit by hand
\name{maHeatmap}
\alias{maHeatmap}
\title{Heatmap of the differentially methylated regions (DMRs) found by a run of methylaction()}
\usage{
maHeatmap(samp, ma, pdf)
}
\arguments{
\item{samp}{Description of samples from readSampleInfo()}

\item{ma}{Output object from methylaction()}

\item{pdf}{PDF file to output.}
}
\value{
Saves plot to disk.
}
\description{
Will plot a heatmap of the DMRs based on the normalized read counts.
}

