% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/reporting.r
\name{maKaryogram}
\alias{maKaryogram}
\title{Karyogram of the differentially methylated regions (DMRs) found by a run of methylaction()}
\usage{
maKaryogram(ma, reads, frequentonly = TRUE, patt = NULL, colors = NULL,
  file = NULL)
}
\arguments{
\item{ma}{Output object from methylaction()}

\item{reads}{Preprocessed reads/fragments data from getReads()}

\item{frequentonly}{Only plot for DMRs where "frequent" is TRUE}

\item{patt}{Character vector of patterns to restrict plot to}

\item{colors}{Character vector of custom colors (as hex codes) for each pattern in patt}

\item{file}{Where to save the image (PDF format), if NULL, will print to current graphics device}
}
\value{
Saves plot to disk or outputs to graphics device
}
\description{
Will plot a karyogram of the DMRs. A black line above shows regions of coverage by the sequencing experiment.
}

