% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/perms.r
\name{maPerm}
\alias{maPerm}
\title{Permute or Bootstrap DMR Detection}
\usage{
maPerm(ma, reads, nperms, combos = NULL, save = T, perm.boot = F,
  ncore = 1)
}
\arguments{
\item{ma}{Output object from methylaction()}

\item{reads}{Preprocessed reads/fragments data from getReads()}

\item{nperms}{Number of permutation (or bootstrap) iterations to perform}

\item{combos}{A matrix of pre-set combinations to use. Useful for smaller sample sizes where there are only a limited number of possible combinations. The row should have ncols = number of samples, and each row represents a re-ordering of samples into groups (where the groups will be set in the order they appear in the "samp" given to methylaction())}

\item{save}{If true, save an RData file of the permutations. These can be merged using maPermMerge(). Useful for running permutations across multiple computers or in a cluster environment. If FALSE, will return the permutation results.}

\item{perm.boot}{If nperms > 0 and if TRUE, perform bootstrapping (sampling with replacement). Otherwise, perform permutations (sampling without permutations)}

\item{ncore}{Number of parallel processes to use}
}
\value{
A list of DMRs arising from each permutation or bootstrap. Will save as an RData if save==TRUE.
}
\description{
Will perform permutations or bootstraps after methylaction() has been called. See also maPermMerge() and maPermFdr().
}

