% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/perms.r
\name{maPermFdr}
\alias{maPermFdr}
\title{Compute false discovery rates (FDRs) from permuted DMR lists}
\usage{
maPermFdr(ma, maperm, recut.p = 0.05)
}
\arguments{
\item{ma}{Output object from methylaction()}

\item{maperm}{Output object from maPerm()}

\item{recut.p}{ANODEV p-value cutoff to use}
}
\value{
A table of FDRs for each pattern, stratified by "frequent" status
}
\description{
Using DMRs arising from permutations saved using maPerm(), compute FDRs for each pattern.
}

