% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/perms.r
\name{maPermMerge}
\alias{maPermMerge}
\title{Merge permutations generated and saved by maPerm()}
\usage{
maPermMerge(dir = ".")
}
\arguments{
\item{dir}{Directory containing RData files saved by maPerm() where save was equal to TRUE}
}
\value{
A list of DMRs arising from each permutation or bootstrap, merged across all the detected RData files.
}
\description{
The function maPerm() with save=TRUE will save permutations into RData files. Move all these RData files into the same directory, and indicate this directory using "dir". The output from maPermMerge() will be a list of DMRs detected in all permutations, and can be given to maPermFdr() to compute false discovery rates (FDRs).
}

