% Generated by roxygen2 (4.0.2): do not edit by hand
\name{maTable}
\alias{maTable}
\title{Table of differentially methylated regions (DMRs) by pattern}
\usage{
maTable(samp, ma)
}
\arguments{
\item{samp}{Description of samples from readSampleInfo()}

\item{ma}{Output object from methylaction()}
}
\value{
A data.frame with the summary statistics.
}
\description{
Will return a table of which patterns were detected and the number of DMRs in each.
}

