% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/reporting.r
\name{maTracks}
\alias{maTracks}
\title{Write BED and BIGWIG files for normalized, filter-passed window count values}
\usage{
maTracks(ma, reads, path = ".", bigwig = FALSE, ncore = 1)
}
\arguments{
\item{ma}{Output list from a run of methylaction()}

\item{reads}{Preprocessed reads/fragments data from getReads()}

\item{path}{Folder to save the files in, will create if does not exist}

\item{bigwig}{If TRUE, convert the BED to BIGWIG files, requires wigToBigWig in $PATH (obtain from the Jim Kent source tree)}

\item{ncore}{Number of parallel processes to use}
}
\value{
Writes BED file to disk and optionally converts to a bigWig file.
}
\description{
Creates a BED file suitable for uploading as a custom track to the UCSC genome browser.
}

