% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/utility.r
\name{readSampleInfo}
\alias{readSampleInfo}
\title{Load a CSV containing required information about each sample}
\usage{
readSampleInfo(file = NULL, colors = NULL)
}
\arguments{
\item{file}{Path to the CSV samplesheet to open. Must contain the columns described above.}

\item{colors}{Vector of colors (one for each group) in same order as groups appear in the sample file. These will be uniform colors used in the plotting functions for these groups. Give colors as hex codes. If none are provided they will be auto-selected using RColorBrewer. Alternatively, if there is a column named "color" in the CSV, then these will always be used.}
}
\value{
A data.frame of the samplesheet that will be valid input to the "samp" arguments of other functions.
}
\description{
The CSV file must contain the following columns: "sample" - unique sample IDs, "group" - group IDs, "bam" - path to BAM file containing aligned reads for the sample. Columns with other names will be retained but ignored. Note that in subsequent reporting of pattern strings (where each digit represents a group), digits for each group will be ordered in the order they first appear in this samplesheet.
}

