% Generated by roxygen2 (4.0.1): do not edit by hand
\name{getCounts}
\alias{getCounts}
\title{Produce a GRanges of windows and their read counts}
\usage{
getCounts(samp, bsgenome, chrs, fragsize, winsize)
}
\arguments{
\item{samp}{Sample data.frame from readSampleInfo()}

\item{begenome}{A b-string genome (bsgenome) object from Bioconductor for the genome of interest.}

\item{chrs}{Which chromosomes to consider.}

\item{fragsize}{Average fragment length from the sequencing experiment. Reads will be extended up to this size when computing coverage.}

\item{winsize}{Size of the non-overlapping windows.}
}
\value{
A GenomicRanges object with values() containing a table of counts for each sample at each window.
}
\description{
Will return a GenomicRanges object for non-overlapping windows genome-wide for the genome given as a bsgenome object for the chromosomes given in chrs. The values() of the GRanges will contain a table of counts for each sample at each window.
}

