% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/utility.r
\name{getCounts}
\alias{getCounts}
\title{Produce a GRanges with counts of overlapping reads for a set of ranges}
\usage{
getCounts(samp, reads, ranges = NULL, chrs = NULL, winsize = 50,
  ncore = 1)
}
\arguments{
\item{samp}{Sample data.frame from readSampleInfo()}

\item{reads}{Output from getReads()}

\item{ranges}{Count within regions given by a GRanges object only}

\item{chrs}{Character vector of chromosome names to select}

\item{winsize}{Size of the non-overlapping windows}

\item{ncore}{Number of parallel processes to use}
}
\value{
A GenomicRanges object with values() containing a table of counts for each sample at each window.
}
\description{
Compute counts in non-overlapping windows genome-wide on the chromosomes given in chrs and return as a GRanges object. The values() of the GRanges will contain a table of counts for each sample at each window.
}

