% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/utility.r
\name{getReads}
\alias{getReads}
\title{Store GenomicRanges of BAM reads in a list}
\usage{
getReads(samp, chrs, fragsize, ncore)
}
\arguments{
\item{samp}{Sample data.frame from readSampleInfo()}

\item{chrs}{Character vector of chromosome names to select}

\item{fragsize}{Average fragment length from the sequencing experiment. Reads will be extended up to this size when computing coverage. If set to the value "paired", valid mate pairs for a paired-end sequencing experiment will be loaded instead.}

\item{ncore}{Number of parallel processes to use}
}
\value{
A list of GenomicRanges objects where the ranges represent either extended reads or mate pairs from a BAM file.
}
\description{
Store positions of BAM reads in a list of GRanges objects
}

