% Generated by roxygen2 (4.0.2): do not edit by hand
\name{maBed}
\alias{maBed}
\title{Write BED file of DMR regions}
\usage{
maBed(ma, file, dmr.only = F)
}
\arguments{
\item{ma}{Output list from a run of methylaction()}

\item{file}{Name of BED file to create}

\item{dmr.only}{Don't report regions without a significant pattern (show up listed as NS, default: FALSE)}
}
\value{
Writes BED file to disk.
}
\description{
Creates a BED file suitable for uploading as a custom track to the UCSC genome browser.
}

