% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/reporting.r
\name{maHeatmap}
\alias{maHeatmap}
\title{Heatmap of the differentially methylated regions (DMRs) found by a run of methylaction()}
\usage{
maHeatmap(ma, frequentonly = TRUE, bias = 2, file = NULL)
}
\arguments{
\item{ma}{Output object from methylaction()}

\item{frequentonly}{Only plot for DMRs where "frequent" is TRUE}

\item{bias}{Bias setting for the color scale}

\item{file}{Where to save the image (PNG format), if NULL, will print to current graphics device}
}
\value{
Saves plot to disk or outputs to graphics device
}
\description{
Will plot a heatmap of the DMRs based on the normalized read counts. The square root of the mean per-window normalized read count is used so DMRs of different lengths are comparable.
}

