% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/reporting.r
\name{maHeatmap}
\alias{maHeatmap}
\title{Heatmap of the differentially methylated regions (DMRs) found by a run of methylaction()}
\usage{
maHeatmap(ma, frequentonly = TRUE, bias = 2, file)
}
\arguments{
\item{ma}{Output object from methylaction()}

\item{samp}{Description of samples from readSampleInfo()}

\item{pdf}{PDF file to output.}
}
\value{
Saves plot to disk.
}
\description{
Will plot a heatmap of the DMRs based on the normalized read counts.
}

