% Generated by roxygen2 (4.0.1): do not edit by hand
\name{maSummary}
\alias{maSummary}
\title{Summary stats for a run of methylaction()}
\usage{
maSummary(samp, ma)
}
\arguments{
\item{samp}{Description of samples from readSampleInfo()}

\item{ma}{Output object from methylaction()}
}
\value{
A data.frame with the summary statistics.
}
\description{
Will return information about number of windows/regions that pass cutoffs at each stage of the analysis. Useful for paramater tuning.
}

