% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/reporting.r
\name{maSummary}
\alias{maSummary}
\title{Summary stats for a run of methylaction()}
\usage{
maSummary(ma)
}
\arguments{
\item{ma}{Output object from methylaction()}
}
\value{
A data.frame with the summary statistics.
}
\description{
Will return information about number of windows/regions that pass cutoffs at each stage of the analysis. Useful for paramater tuning.
}

