% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/methylaction-package.r
\docType{package}
\name{methylaction-package}
\alias{methylaction-package}
\title{methylaction: Differentially Methylated Regions (DMRs) from MBD-isolated Genome Sequencing (MiGS/MBD-seq)}
\description{
methylaction: Differentially Methylated Regions (DMRs) from MBD-isolated Genome Sequencing (MiGS/MBD-seq)
}

