% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{package}
\name{methylaction-package}
\alias{methylaction-package}
\title{methylaction: Differentially Methylated Regions (DMRs) from MBD-isolated Genome Sequencing (MiGS/MBD-seq)}
\description{
methylaction: Differentially Methylated Regions (DMRs) from MBD-isolated Genome Sequencing (MiGS/MBD-seq)
}

