% Generated by roxygen2 (4.0.2): do not edit by hand
\name{methylaction}
\alias{methylaction}
\title{Detect differentially methylated regions (DMRs) from windowed read counts from MBD-isolated genome sequencing (MiGS/MBD-seq)}
\usage{
methylaction(samp, counts, reads, winsize, poifdr, stageone.p, joindist,
  anodev.p, post.p, minsize = 150, nperms = 0, ncore = 1)
}
\arguments{
\item{samp}{Description of samples from readSampleInfo()}

\item{counts}{Preprocessed count data from getCounts()}

\item{winsize}{Size of the windows used when counting.}

\item{poifdr}{False discovery rate to use during initital filtering.}

\item{stageone.p}{P-value cutoff for stage one testing.}

\item{anodev.p}{P-value cutoff for the analysis of deviance (ANODEV) in stage two testing (ignored for two group comparisions).}

\item{post.p}{P-value cutoff for post-tests (or for the single test stage two test in the two group case).}

\item{minsize}{Minimum size for a reported region.}

\item{ncore}{Number of cores to use.}

\item{bsgenome}{b-string genome (bsgenome) object for the genome}

\item{fragsize}{The average fragment length selected for in the sequencing experiment (used to extend reads when re-counting for regions in stage two testing).}
}
\value{
A list containing detailed results from each stage of the analysis.
}
\description{
Once the counts have been pre-processed, this function performs all the analysis. Detailed results from intermediate steps are stored in the output list object to analyze method performance and provide input for the summary and plotting functions.
}

